requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
django-environ==0.8.1

gunicorn==20.1.0 

google-cloud-secret-manager==2.10.0
django-storages[google]==1.12.3

mysqlclient==2.1.0

xendit==0.1.3
xendit-python==0.1.4

asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
idna==3.3
iniconfig==1.1.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==2.12.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.8
whitenoise==6.0.0